refactor(sumcheck): rename ring_switch's pack module to packing
ring_switch::pack names both a module and, via re-export, a function,
which reviewers flagged as confusing at call sites and in docs. The
module becomes packing; the pack function keeps its name and its
re-export at the ring_switch root.
